+++ src/app/lib/lforms-widget-registry.ts
@@ -65,6 +65,7 @@ import {
   CalculatedInitialExpressionComponent
 } from "./widgets/expression-editor/calculated-initial-expression/calculated-initial-expression.component";
 import {EnableWhenExpressionComponent} from "./widgets/expression-editor/enable-when-expression/enable-when-expression.component";
+import {MinMaxOccursComponent} from './widgets/min-max-occurs/min-max-occurs.component';
 
 @Injectable()
 export class LformsWidgetRegistry extends DefaultWidgetRegistry {
@@ -128,6 +129,7 @@ export class LformsWidgetRegistry extends DefaultWidgetRegistry {
     this.register('coding-display', CodingDisplayComponent);
     this.register('enable-when-method', EnableWhenMethodComponent);
     this.register('enable-when-expression', EnableWhenExpressionComponent);
+    this.register('min-max-occurs', MinMaxOccursComponent);
 
     this.setDefaultWidget(StringComponent);
   }