+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This project follows [Semantic Versioning](http://semver.org/).
 
+## [11.2.0] 2026-04-27
+### Added
+- Added support for `questionnaire-minOccurs` and `questionnaire-maxOccurs` extensions
+  via a combined Min/Max Occurs widget that is shown when an item allows repeats.
+
 ## [11.1.0] 2026-04-23
 ### Added
 - Added support to create general FHIR extensions including nested extensions.